Added GET docs/:track endpoint for serving language-specific documentation
This endpoint is used for serving language-specific documentation.
This endpoint render a json object whose value of "docs" is a json object with the documentation
filenames lower-cased without extension (e.g. "TOOLS.md" -> "tools") as keys and contents as values.
If the track does not exists or the track does not have documentation, it returns a 404 error.