Mark individual exercises as fetched
When someone fetches the next exercise (whichever it is) either
across all their active tracks or in a specific track, we post
back to the app so that we can show people their current status.

When someone fetches an individual exercise, we weren't posting back.
This fixes it for the individual fetch.

This does presume that the CLI passes along the API key, which it
originally was not doing. That's fixed in
https://github.com/exercism/cli/pull/307
and it will be a part of the 2.2.6 release.