Rewrite implementation of restore endpoint
We were intermittently getting very, very strange behavior in
production, which was very, very difficult to debug.

Some people were getting a (seemingly random) solution delivered
along with the exercise implementation. It was not the
solution that came from the example.rb file, it was some other user's
solution.

It turns out that the version of Ruby that is running on Heroku has
a different behavior for `dup` than the version running locally
on my machine.

We were calling dup on a cached exercise implementation, then mutating
the referenced "files" hash in the dup (not the original). However,
on Heroku, dup creates an object that has a reference to the same
hash, rather than also copying the hash, causing the original hash
to be mutated. The behavior on my local machine is creating a copy of
the hash, and so this behavior is not reproducible locally.

To test both locally and on heroku I verified the object_ids of the
files attribute of an implementation.

    Xapi.tracks['ruby'].implementations['gigasecond'].dup.files.object_id
    Xapi.tracks['ruby'].implementations['gigasecond'].files.object_id

Heroku had the same object_id, my local version did not.

Fixes exercism/exercism.io#2882
Fixes exercism/xruby#315