Fix non-deterministic file ordering.
The order in which files were returned by Dir.glob is dependent on the
file ordering in the underlying filesystem. This was causing unnecessary
test failures.

The solution was to make sure all relevant directory reads return a
sorted array of filenames.

Also:

100% test coverage for /lib/xapi/track.rb in test/xapi/track_test.rb
Updated fixtures that needed updating.
Removed redundant tests of Ruby's regex engine.