Report most popular track documentation format.
The API reports the format of files in the docs directory based on file
extension. Theoretically these should all be the same.

The previous implementation would use the format of the first file in
the directory, This version uses the most popular extension.

The format exension is currently only(?) used to work out what the name
of the file to edit is at the bottom of the track documentation on the
exercism.io site.