Handle new metadata file structure
Previously we've had all the exercise metadata files in the root folder.

```
.
├── acronym.json
├── acronym.md
└── acronym.yml
```

We are currently in the process of rearranging these to be nested under
directories named for the exercise:

```
.
└── exercises
    └── acronym
        ├── canonical-data.json
        ├── description.md
        └── metadata.yml
```

This change supports both formats, defaulting to the new one.

Once the changes have been merged into the x-common directory, we
can delete the support for the old structure.