Eliminate implementation mutation during restore.
Mutating the implementation files was causing problems.

This patch returns the files to be restored without mutating the
original implementation object.

A Restoration class which can be used to construct the restorations we
need has been extracted.

A restoration is a combination of an implementation plus the files that
implementation needs to be restored, which is the implementation files
plus the most recent solution file.

This is then passed to the restore petroglyph view for rendering.